Refactor comments presenter to a view component
We've decided to use view components in favor of presenters.

Fixes #11559
