+++ src/api/app/components/comment_component.xml.builder
@@ -1,0 +1,3 @@
+@builder.comment_(attributes) do
+  @builder.text(body)
+end
