--- src/api/app/views/comments/_comment.xml.builder
@@ -1,5 +1,0 @@
-comment_presenter = CommentsControllerPresenters::CommentPresenter.new(comment, obj_is_user)
-
-builder.comment_(comment_presenter.attributes) do
-  builder.text(comment_presenter.body)
-end
