+++ src/api/app/views/comments/index.xml.builder
@@ -1,4 +1,3 @@
 xml.comments(@header) do
-  render(partial: 'comment', collection: @comments, locals: { obj_is_user: @obj.is_a?(User),
-                                                              builder: xml })
+  render(CommentComponent.with_collection(@comments, obj_is_user: @obj.is_a?(User), builder: xml))
 end
